requirements.txt
textual~=8.1.1
rich~=14.3.3
google-genai
python-dotenv
protobuf